refactor: UserDao와 ConnectionMaker 간 의존 관계 설정 책임의 분리
