+++ src/main/java/springbook/dao/UserDaoTest.java
@@ -1,13 +1,13 @@
-package springbook;
+package springbook.dao;
 
-import springbook.dao.UserDao;
 import springbook.domain.User;
 
 import java.sql.SQLException;
 
-public class Application {
+public class UserDaoTest {
     public static void main(String[] args) throws ClassNotFoundException, SQLException {
-        UserDao dao = new UserDao();
+        ConnectionMaker connectionMaker = new SimpleConnectionMaker();
+        UserDao dao = new UserDao(connectionMaker);
 
         User user = new User();
         user.setId("whiteship");
